PTERO-T-7 Fixed ELUV method behaviour and added tests
